split out BitstreamController and SeenFrame from SimulcastController
break out SeenFrame from BitstreamController

slight tweak to the filtering logic in SeenFrame#accept

add unit test for SeenFrame#accept

BitstreamController unit tests

update license headers for new test files

wip: simulcastcontroller unit tests

add authors
