+++ pom.xml
@@ -191,6 +191,20 @@
       <version>3.4</version>
       <scope>test</scope>
     </dependency>
+	<!-- https://mvnrepository.com/artifact/org.powermock/powermock-module-junit4 -->
+	<dependency>
+		<groupId>org.powermock</groupId>
+		<artifactId>powermock-module-junit4</artifactId>
+		<version>1.7.0</version>
+		<scope>test</scope>
+	</dependency>
+    <!-- https://mvnrepository.com/artifact/org.powermock/powermock-api-easymock -->
+	<dependency>
+		<groupId>org.powermock</groupId>
+		<artifactId>powermock-api-easymock</artifactId>
+		<version>1.7.0</version>
+		<scope>test</scope>
+	</dependency>
   </dependencies>
 
   <build>
