+++ src/main/java/org/jitsi/videobridge/cc/SimulcastController.java
@@ -17,8 +17,6 @@ package org.jitsi.videobridge.cc;
 
 import net.sf.fmj.media.rtp.*;
 import org.jitsi.impl.neomedia.*;
-import org.jitsi.impl.neomedia.codec.*;
-import org.jitsi.impl.neomedia.codec.video.vp8.*;
 import org.jitsi.impl.neomedia.rtcp.*;
 import org.jitsi.impl.neomedia.rtp.*;
 import org.jitsi.impl.neomedia.rtp.translator.*;
@@ -26,12 +24,9 @@ import org.jitsi.impl.neomedia.transform.*;
 import org.jitsi.service.configuration.*;
 import org.jitsi.service.libjitsi.*;
 import org.jitsi.service.neomedia.*;
-import org.jitsi.service.neomedia.codec.*;
 import org.jitsi.util.*;
-import org.jitsi.util.function.*;
 
 import java.lang.ref.*;
-import java.util.*;
 
 /**
  * Filters the packets coming from a specific {@link MediaStreamTrackDesc}
@@ -46,41 +41,10 @@ public class SimulcastController
     implements AutoCloseable
 {
     /**
-     * The property name of the number of seen frames to keep track of.
-     */
-    public static final String SEEN_FRAME_HISTORY_SIZE_PNAME
-        = "org.jitsi.videobridge.SEEN_FRAME_HISTORY_SIZE";
-
-    /**
-     *
-     */
-    public static final String ENABLE_VP8_PICID_REWRITING_PNAME
-        = "org.jitsi.videobridge.ENABLE_VP8_PICID_REWRITING";
-
-    /**
      * The ConfigurationService to get config values from.
      */
     private static final ConfigurationService
         cfg = LibJitsi.getConfigurationService();
-
-    /**
-     * The default number of seen frames to keep track of.
-     */
-    private static final int SEEN_FRAME_HISTORY_SIZE_DEFAULT = 120;
-
-    /**
-     * The number of seen frames to keep track of.
-     */
-    private static final int SEEN_FRAME_HISTORY_SIZE =
-        cfg != null ? cfg.getInt(SEEN_FRAME_HISTORY_SIZE_PNAME,
-            SEEN_FRAME_HISTORY_SIZE_DEFAULT) : SEEN_FRAME_HISTORY_SIZE_DEFAULT;
-
-    /**
-     * A boolean that indicates whether or not to activate VP8 picture ID
-     * rewriting.
-     */
-    private static final boolean ENABLE_VP8_PICID_REWRITING =
-        cfg != null && cfg.getBoolean(ENABLE_VP8_PICID_REWRITING_PNAME, false);
 
     /**
      * Index used to represent that forwarding is suspended.
@@ -111,15 +75,6 @@ public class SimulcastController
     private final long targetSsrc;
 
     /**
-     * The running index for the temporal base layer frames, i.e., the frames
-     * with TID set to 0. Introduced for VP8 PictureID rewriting. The initial
-     * value is chosen so that it matches the TL0PICIDX of the black VP8 key
-     * frames injected by the lipsync hack; so the code assumes a packet of TID
-     * 0 has already been sent.
-     */
-    private int tl0PicIdx = 0;
-
-    /**
      * The {@link BitstreamController} for the currently forwarded RTP stream.
      */
     private final BitstreamController bitstreamController;
@@ -132,7 +87,8 @@ public class SimulcastController
      * @param source the source {@link MediaStreamTrackDesc}
      */
     SimulcastController(
-        BitrateController bitrateController, MediaStreamTrackDesc source)
+        BitrateController bitrateController, MediaStreamTrackDesc source,
+        BitstreamController bitstreamController)
     {
         this.bitrateController = bitrateController;
         weakSource = new WeakReference<>(source);
@@ -147,7 +103,7 @@ public class SimulcastController
             targetSsrc = rtpEncodings[0].getPrimarySSRC();
         }
 
-        bitstreamController = new BitstreamController();
+        this.bitstreamController = bitstreamController;
     }
 
     /**
@@ -269,7 +225,8 @@ public class SimulcastController
             // Regardless of whether a switch is pending or not, if an incoming
             // frame belongs to the current stream being forwarded, we'll
             // accept it (if the bitstreamController lets it through)
-            return bitstreamController.accept(sourceFrameDesc, pkt);
+            return bitstreamController.accept(sourceFrameDesc, pkt,
+                bitrateController.getVideoChannel().getStream());
         }
 
         // At this point we know that we want to be forwarding something and
@@ -290,8 +247,11 @@ public class SimulcastController
             {
                 // This frame represents, at least, a step in the right direction
                 // towards the stream we want
-                bitstreamController.setTL0Idx(sourceBaseLayerIndex);
-                return bitstreamController.accept(sourceFrameDesc, pkt);
+                MediaStreamTrackDesc source = weakSource.get();
+                assert source != null;
+                bitstreamController.setTL0Idx(sourceBaseLayerIndex, source.getRTPEncodings());
+                return bitstreamController.accept(sourceFrameDesc, pkt,
+                    bitrateController.getVideoChannel().getStream());
             }
         }
         return false;
@@ -402,7 +362,11 @@ public class SimulcastController
             return new RawPacket[] { pktIn };
         }
 
-        RawPacket[] pktsOut = bitstreamController.rtpTransform(pktIn);
+        MediaStreamTrackDesc source = weakSource.get();
+        assert source != null;
+        RawPacket[] pktsOut =
+            bitstreamController.rtpTransform(pktIn,
+                source.getMediaStreamTrackReceiver().getStream());
 
         if (!ArrayUtils.isNullOrEmpty(pktsOut)
             && pktIn.getSSRCAsLong() != targetSsrc)
@@ -510,1065 +474,6 @@ public class SimulcastController
     public void close()
         throws Exception
     {
-        bitstreamController.setTL0Idx(-1);
-    }
-
-    class BitstreamController
-    {
-        /**
-         * The available subjective quality indexes that this RTP stream offers.
-         * NOTE: this refers to the available qualities for the current stream
-         * only.  a non-adaptive stream would only have 1, an adaptive stream
-         * would have multiple for each of its layers.
-         */
-        private int[] availableQualityIndices;
-
-        /**
-         * A boolean that indicates whether or not the current TL0 is adaptive
-         * (meaning it has multiple quality layers within the same rtp stream) or
-         * not.
-         */
-        private boolean isAdaptive;
-
-        /**
-         * The sequence number offset that this bitstream started. The initial
-         * value -1 is an indication to use the first accepted packet seqnum
-         * as the offset.
-         */
-        private int seqNumOff = -1;
-
-        /**
-         * The timestamp offset that this bitstream started. The initial
-         * value -1 is an indication to use the first accepted packet timestamp
-         * as the offset.
-         */
-        private long tsOff = -1;
-
-        /**
-         * The running index offset of the frames that are sent out. Introduced
-         * for VP8 PictureID rewriting. The initial value is chosen so that it
-         * matches the pictureID of the black VP8 key frames injected by the
-         * lipsync hack, so the code assumes a packet with picture ID 1 has
-         * already been sent.
-         */
-        private int pidOff = 1;
-
-        /**
-         * The running index delta to apply to the frames that are sent out.
-         * Introduced for VP8 PictureID rewriting. The initial value is chosen
-         * so that it matches the pictureID of the black VP8 key frames injected
-         * by the lipsync hack, so the code assumes a packet with picture ID 1
-         * has already been sent.
-         */
-        private int pidDelta = -1;
-
-        /**
-         * The SSRC of the TL0 of the RTP stream that is currently being
-         * forwarded. This is useful for simulcast and RTCP SR rewriting.
-         *
-         * The special value -1 indicates that we're not accepting (and hence
-         * not forwarding) anything.
-         */
-        private long tl0Ssrc = -1;
-
-        /**
-         * The subjective quality index for of the TL0 of this instance.
-         *
-         * The special value -1 indicates that we're not accepting (and hence
-         * not forwarding) anything.
-         */
-        private int tl0Idx = -1;
-
-        /**
-         * The target subjective quality index for this instance. This instance
-         * switches between the available RTP streams and sub-encodings until it
-         * reaches this target. -1 effectively means that the stream is
-         * suspended.
-         */
-        private int targetIdx = -1;
-
-        /**
-         * The optimal subjective quality index for this instance.
-         */
-        private int optimalIdx = -1;
-
-        /**
-         * The current subjective quality index for this instance. If this is
-         * different than the target, then a switch is pending.
-         *
-         * When SVC is enabled, currentIdx >= tl0Idx, otherwise
-         * currentIdx == tl0Idx.
-         */
-        private int currentIdx = -1;
-
-        /**
-         * The number of transmitted bytes.
-         */
-        private long transmittedBytes = 0;
-
-        /**
-         * The number of transmitted packets.
-         */
-        private long transmittedPackets = 0;
-
-        /**
-         * The most recent key frame that we've sent out. Anything that's
-         * accepted by this controller needs to be newer than this.
-         */
-        private SeenFrame mostRecentSentKeyFrame;
-
-        /**
-         * The max (biggest timestamp) frame that we've sent out.
-         */
-        private SeenFrame mostRecentSentFrame;
-
-        /**
-         * The {@link SeenFrameAllocator} for this {@link BitstreamController}.
-         */
-        private final SeenFrameAllocator seenFrameAllocator
-            = new SeenFrameAllocator();
-
-        /**
-         * At 60fps, this holds 2 seconds worth of frames.
-         * At 30fps, this holds 4 seconds worth of frames.
-         */
-        private final Map<Long, SeenFrame> seenFrames
-            = Collections.synchronizedMap(
-                new LRUCache<Long, SeenFrame>(SEEN_FRAME_HISTORY_SIZE)
-        {
-            /**
-             * {@inheritDoc}
-             */
-            protected boolean removeEldestEntry(Map.Entry<Long, SeenFrame> eldest)
-            {
-                boolean removeEldestEntry = super.removeEldestEntry(eldest);
-
-                if (removeEldestEntry)
-                {
-                    seenFrameAllocator.returnSeenFrame(eldest.getValue());
-                }
-
-                return removeEldestEntry;
-            }
-        });
-
-        public void suspend()
-        {
-            setTL0Idx(-1);
-        }
-
-        /**
-         *
-         * @param newTL0Idx
-         */
-        void setTL0Idx(int newTL0Idx)
-        {
-            if (this.tl0Idx == newTL0Idx)
-            {
-                return;
-            }
-
-            if (logger.isDebugEnabled())
-            {
-                logger.debug("tl0_changed,hash="
-                    + SimulcastController.this.hashCode()
-                    + " old_tl0=" + this.tl0Idx
-                    + ",new_tl0=" + newTL0Idx);
-            }
-
-            this.seenFrames.clear();
-
-            this.mostRecentSentKeyFrame = null;
-            if (mostRecentSentFrame != null)
-            {
-                this.tsOff = getMaxTs();
-                this.seqNumOff = getMaxSeqNum();
-                this.pidOff = getMaxPictureID();
-                this.mostRecentSentFrame = null;
-            }
-
-            this.pidDelta = -1;
-
-            int oldTL0Idx = this.tl0Idx;
-            this.tl0Idx = newTL0Idx;
-
-            // a stream always starts suspended (and resumes with a key frame).
-            this.currentIdx = -1;
-
-            MediaStreamTrackDesc source = weakSource.get();
-            assert source != null;
-            RTPEncodingDesc[] rtpEncodings = source.getRTPEncodings();
-            if (!ArrayUtils.isNullOrEmpty(rtpEncodings))
-            {
-                if (oldTL0Idx > -1)
-                {
-                    rtpEncodings[rtpEncodings[oldTL0Idx].getBaseLayer().getIndex()].decrReceivers();
-                }
-
-                if (newTL0Idx > -1)
-                {
-                    rtpEncodings[rtpEncodings[newTL0Idx].getBaseLayer().getIndex()].incrReceivers();
-                }
-            }
-            if (ArrayUtils.isNullOrEmpty(rtpEncodings) || tl0Idx < 0)
-            {
-                this.availableQualityIndices = null;
-                this.tl0Ssrc = -1;
-                this.isAdaptive = false;
-            }
-            else
-            {
-                tl0Idx = rtpEncodings[tl0Idx].getBaseLayer().getIndex();
-                tl0Ssrc = rtpEncodings[tl0Idx].getPrimarySSRC();
-
-                // find the available qualities in this bitstream.
-
-                // TODO optimize if we have a single quality
-                List<Integer> availableQualities = new ArrayList<>();
-                for (int i = 0; i < rtpEncodings.length; i++)
-                {
-                    if (rtpEncodings[i].requires(tl0Idx))
-                    {
-                        availableQualities.add(i);
-                    }
-                }
-
-                availableQualityIndices = new int[availableQualities.size()];
-                Iterator<Integer> iterator = availableQualities.iterator();
-                for (int i = 0; i < availableQualityIndices.length; i++)
-                {
-                    availableQualityIndices[i] = iterator.next();
-                }
-
-                this.isAdaptive = availableQualityIndices.length > 1;
-            }
-        }
-
-        /**
-         * Given a list of available quality indices for the current stream
-         * and a target index, find the highest available quality index <= targetIdx.
-         * Note that targetIdx may be pointing at a different stream, i.e.
-         * availableQualityIndices may be {7, 8, 9} and targetIdx could be 3.
-         * We won't change the stream we're sending, but will find the max index
-         * that is both in the current stream and is less than the targetIdx,
-         * if it exists
-         * @param availableQualityIndices the available quality indices
-         * @param targetIdx the target index
-         * @return the highest quality index <= targetIdx
-         * //FIXME(brian): this could be static, but can't add a static method
-         * to the inner class.  should it be put elsewhere and made a static
-         * helper method? (the outer class?)
-         */
-        private int findMaximumQualityIndex(int[] availableQualityIndices, int targetIdx)
-        {
-            int maxIndex = -1;
-            if (availableQualityIndices != null && availableQualityIndices.length > 0)
-            {
-                // At the very least we will send the lowest quality layer of
-                // the current stream.  If there happens to be an additional,
-                // higher index of this stream that is also <= targetIdx, we will
-                // upgrade to that
-                maxIndex = availableQualityIndices[0];
-                for (int i = 1; i < availableQualityIndices.length; ++i)
-                {
-                    if (availableQualityIndices[i] <= targetIdx)
-                    {
-                        maxIndex = availableQualityIndices[i];
-                    }
-                    else
-                    {
-                        break;
-                    }
-                }
-
-            }
-            return maxIndex;
-        }
-
-        /**
-         * Defines an RTP packet filter that controls which packets to be written
-         * into some arbitrary target/receiver that owns this
-         * {@link BitstreamController}.
-         *
-         * @param sourceFrameDesc the {@link FrameDesc} that the RTP packet belongs
-         * to.
-         * @param pkt the packet for which to decide to accept
-         * @return <tt>true</tt> to allow the specified packet/<tt>buffer</tt> to be
-         * written into the arbitrary target/receiver that owns this
-         * {@link BitstreamController} ; otherwise, <tt>false</tt>
-         */
-        public boolean accept(
-            FrameDesc sourceFrameDesc, RawPacket pkt)
-        {
-            if (isAdaptive && !sourceFrameDesc.firstSequenceNumberKnown())
-            {
-                return false;
-            }
-
-            long srcTs = sourceFrameDesc.getTimestamp();
-            SeenFrame destFrame = seenFrames.get(srcTs);
-
-            if (destFrame == null)
-            {
-                // An unseen frame has arrived. We forward it iff all of the
-                // following conditions hold:
-                //
-                // 1. it's a dependency of the currentIdx (obviously).
-                // 2. it's newer than max frame (we can't modify the distances of
-                // past frames).
-                // 2. we know the boundaries of the max frame OR if the max frame is
-                // *not* a TL0 (this allows for non-TL0 to be corrupt).
-                //
-                // Given the above conditions, we might decide to drop a TL0 frame.
-                // This can happen when the max frame is a TL0 and we don't know its
-                // boundaries. Then the stream will be broken an we should ask for a
-                // key frame.
-                {
-                    int newCurrentIdx = findMaximumQualityIndex(availableQualityIndices, targetIdx);
-                    if (newCurrentIdx != this.currentIdx && logger.isDebugEnabled())
-                    {
-                        logger.debug("current_idx_changed,hash="
-                            + SimulcastController.this.hashCode()
-                            + " old_idx=" + this.currentIdx
-                            + ",new_idx=" + newCurrentIdx);
-                    }
-                    this.currentIdx = newCurrentIdx;
-                }
-
-                boolean isNewerThanMostRecentFrame = !haveSentFrame()
-                    || RTPUtils.isNewerTimestampThan(srcTs, mostRecentSentFrame.srcTs);
-
-                boolean isNewerThanMostRecentKeyFrame = !haveSentKeyFrame()
-                    || RTPUtils.isNewerTimestampThan(srcTs, mostRecentSentKeyFrame.srcTs);
-
-                if (!isSuspended()
-                    // we haven't seen anything yet and this is an independent
-                    // frame.
-                    && (!haveSentFrame() && sourceFrameDesc.isIndependent()
-                    // frames from non-adaptive streams need to be newer than
-                    // the most recent independent frame
-                    || (haveSentFrame() && !isAdaptive
-                        && isNewerThanMostRecentKeyFrame)
-                    // frames from adaptive streams need to be newer than the
-                    // max
-                    || (haveSentFrame() && isAdaptive && isNewerThanMostRecentFrame)))
-                {
-                    // the stream is not suspended and we're not dealing with a
-                    // late frame or the stream is not adaptive.
-
-                    RTPEncodingDesc sourceEncodings[] = sourceFrameDesc
-                        .getRTPEncoding().getMediaStreamTrack().getRTPEncodings();
-
-                    // TODO ask for independent frame if we're skipping a TL0.
-
-                    int sourceIdx = sourceFrameDesc.getRTPEncoding().getIndex();
-                    if (sourceEncodings[this.currentIdx].requires(sourceIdx))
-                    {
-                        // the quality of the frame is a dependency of the
-                        // forwarded quality and the max frame is effectively
-                        // complete.
-
-                        SeqNumTranslation seqNumTranslation;
-                        if (!haveSentFrame() || isAdaptive)
-                        {
-                            int maxSeqNum = getMaxSeqNum();
-                            if (maxSeqNum > -1)
-                            {
-                                int seqNumDelta = (maxSeqNum + 1
-                                    - sourceFrameDesc.getStart()) & 0xFFFF;
-
-                                seqNumTranslation
-                                    = new SeqNumTranslation(seqNumDelta);
-                            }
-                            else
-                            {
-                                seqNumTranslation = null;
-                            }
-                        }
-                        else
-                        {
-                            // If the current bitstream only offers a single
-                            // quality, then we're not filtering anything thus
-                            // the sequence number distances between the frames
-                            // are fixed so we can reuse the sequence number
-                            // translation from the mostRecentSentFrame.
-                            seqNumTranslation = mostRecentSentFrame.seqNumTranslation;
-                        }
-
-                        TimestampTranslation tsTranslation;
-                        if (!haveSentFrame())
-                        {
-                            if (tsOff > -1)
-                            {
-                                long tsDelta = (tsOff + 3000 - srcTs) & 0xFFFFFFFFL;
-
-                                tsTranslation = new TimestampTranslation(tsDelta);
-                            }
-                            else
-                            {
-                                tsTranslation = null;
-                            }
-                        }
-                        else
-                        {
-                            // The timestamp delta doesn't change for a bitstream,
-                            // so, if we've sent out a frame already, reuse its
-                            // timestamp translation.
-                            tsTranslation = mostRecentSentFrame.tsTranslation;
-                        }
-
-                        int dstPictureID = -1, dstTL0PICIDX = -1;
-                        if (ENABLE_VP8_PICID_REWRITING)
-                        {
-                            byte vp8PT = bitrateController.getVideoChannel()
-                                .getStream().getDynamicRTPPayloadType(Constants.VP8);
-
-                            boolean isVP8 = vp8PT == pkt.getPayloadType();
-
-                            if (isVP8)
-                            {
-                                dstPictureID = calculatePictureID(pkt);
-                            }
-
-                            int tid = ((MediaStreamImpl) bitrateController
-                                .getVideoChannel().getStream())
-                                .getTemporalID(pkt);
-
-                            if (tid > -1)
-                            {
-                                if (tid == 0)
-                                {
-                                    tl0PicIdx++;
-                                }
-
-                                dstTL0PICIDX = tl0PicIdx;
-                            }
-                        }
-
-                        destFrame = seenFrameAllocator.getOrCreate();
-                        destFrame.reset(srcTs, seqNumTranslation, tsTranslation,
-                            dstPictureID, dstTL0PICIDX, sourceFrameDesc.isIndependent());
-                        seenFrames.put(srcTs, destFrame);
-
-                        if (isNewerThanMostRecentFrame)
-                        {
-                            mostRecentSentFrame = destFrame;
-                        }
-
-                        if (isNewerThanMostRecentKeyFrame
-                            && sourceFrameDesc.isIndependent())
-                        {
-                            mostRecentSentKeyFrame = destFrame;
-                        }
-                    }
-                    else
-                    {
-                        destFrame = seenFrameAllocator.getOrCreate();
-                        destFrame.reset(srcTs);
-                        seenFrames.put(srcTs, destFrame);
-                    }
-                }
-                else
-                {
-                    // TODO ask for independent frame if we're filtering a TL0.
-
-                    destFrame = seenFrameAllocator.getOrCreate();
-                    destFrame.reset(srcTs);
-                    seenFrames.put(srcTs, destFrame);
-                }
-            }
-
-            boolean accept = destFrame.accept(sourceFrameDesc, pkt);
-
-            if (accept)
-            {
-                transmittedPackets++;
-                transmittedBytes += pkt.getLength();
-            }
-
-            return accept;
-        }
-
-        /**
-         * Calculates the destination picture ID of an incoming VP8 frame.
-         *
-         * @param pkt the packet from which to calculate the vp8 picture id
-         * @return the destination picture ID of the VP8 frame that is specified
-         * in the parameter.
-         * FIXME(brian): feels like this should be in some vp8-specific class
-         */
-        private int calculatePictureID(RawPacket pkt)
-        {
-            REDBlock redBlock = bitrateController.getVideoChannel()
-                    .getStream().getPrimaryREDBlock(pkt);
-
-            int srcPID = DePacketizer
-                .VP8PayloadDescriptor.getPictureId(
-                    redBlock.getBuffer(),
-                    redBlock.getOffset());
-
-            int dstPictureID;
-            if (srcPID > -1)
-            {
-                if (pidDelta == -1)
-                {
-                    pidDelta = (pidOff + 1 - srcPID) & 0x7FFF;
-                }
-
-                dstPictureID = (srcPID + pidDelta) & 0x7FFF;
-
-                if (((dstPictureID - getMaxPictureID()) & 0x7FFF) > 200
-                    || ((dstPictureID - getMaxPictureID()) & 0x7FFF) > 0x7F00)
-                {
-                    pidDelta = (getMaxPictureID() + 1 - srcPID) & 0x7FFF;
-                    dstPictureID = (srcPID + pidDelta) & 0x7FFF;
-                    logger.warn("A jump was detected in the picture IDs.");
-                }
-            }
-            else
-            {
-                dstPictureID = -1;
-            }
-
-            return dstPictureID;
-        }
-
-        /**
-         * Gets the optimal subjective quality index for this instance.
-         */
-        int getOptimalIndex()
-        {
-            return optimalIdx;
-        }
-
-        /**
-         * Gets the current subjective quality index for this instance.
-         */
-        int getCurrentIndex()
-        {
-            return currentIdx;
-        }
-
-        /**
-         * Gets the target subjective quality index for this instance.
-         */
-        int getTargetIndex()
-        {
-            return targetIdx;
-        }
-
-        /**
-         * Sets the target subjective quality index for this instance.
-         *
-         * @param newTargetIdx the new target subjective quality index for this
-         * instance.
-         */
-        void setTargetIndex(int newTargetIdx)
-        {
-            this.targetIdx = newTargetIdx;
-        }
-
-        /**
-         * Sets the optimal subjective quality index for this instance.
-         *
-         * @param newOptimalIdx the new optimal subjective quality index for this
-         * instance.
-         */
-        void setOptimalIndex(int newOptimalIdx)
-        {
-            this.optimalIdx = newOptimalIdx;
-        }
-
-        /**
-         * Translates accepted packets and drops packets that are not accepted (by
-         * this instance).
-         *
-         * @param pktIn the packet to translate or drop.
-         *
-         * @return the translated packet (along with any piggy backed packets), if
-         * the packet is accepted, null otherwise.
-         */
-        RawPacket[] rtpTransform(RawPacket pktIn)
-        {
-            if (pktIn.getSSRCAsLong() != tl0Ssrc)
-            {
-                return null;
-            }
-
-            SeenFrame destFrame = seenFrames.get(pktIn.getTimestamp());
-            if (destFrame == null)
-            {
-                return null;
-            }
-
-            return destFrame.rtpTransform(pktIn);
-        }
-
-        /**
-         * Updates the timestamp, transmitted bytes and transmitted packets in the
-         * RTCP SR packets.
-         */
-        void rtcpTransform(ByteArrayBuffer baf)
-        {
-            // Make a local reference to this.maxSentFrame to prevent
-            // it from being set to null or a different frame while we're
-            // using it.  maxSentFrame.tsTranslation is conceptually final for
-            // this given frame, so we don't need any extra protection
-            SeenFrame localMaxSentFrameCopy = this.mostRecentSentFrame;
-            // Rewrite timestamp.
-            if (localMaxSentFrameCopy != null && localMaxSentFrameCopy.tsTranslation != null)
-            {
-                long srcTs = RTCPSenderInfoUtils.getTimestamp(baf);
-                long dstTs = localMaxSentFrameCopy.tsTranslation.apply(srcTs);
-
-                if (srcTs != dstTs)
-                {
-                    RTCPSenderInfoUtils.setTimestamp(baf, (int) dstTs);
-                }
-            }
-
-            // Rewrite packet/octet count.
-            RTCPSenderInfoUtils.setOctetCount(baf, (int) transmittedBytes);
-            RTCPSenderInfoUtils.setPacketCount(baf, (int) transmittedPackets);
-        }
-
-        /**
-         * Gets the maximum sequence number that this instance has accepted.
-         */
-        private int getMaxSeqNum()
-        {
-            return haveSentFrame() ? mostRecentSentFrame.getMaxSeqNum() : seqNumOff;
-        }
-
-        /**
-         * Gets the maximum VP8 picture ID that this instance has accepted.
-         */
-        private int getMaxPictureID()
-        {
-            return haveSentFrame() ? mostRecentSentFrame.dstPictureID : pidOff;
-        }
-
-        /**
-         * Gets the maximum timestamp that this instance has accepted.
-         */
-        private long getMaxTs()
-        {
-            return haveSentFrame() ? mostRecentSentFrame.getTs() : tsOff;
-        }
-
-        /**
-         * Gets the SSRC of the TL0 of the RTP stream that is currently being
-         * forwarded.
-         *
-         * @return the SSRC of the TL0 of the RTP stream that is currently being
-         * forwarded.
-         */
-        long getTL0SSRC()
-        {
-            return tl0Ssrc;
-        }
-
-        /**
-         * Return whether or not this BitstreamController has sent a frame
-         * from the current bitstream yet
-         *
-         * @return true if it has sent a frame from this bitstream, false
-         * otherwise
-         */
-        private boolean haveSentFrame()
-        {
-            return mostRecentSentFrame != null;
-        }
-
-        /**
-         * Return whether or not this BitstreamController has sent a key frame
-         * from the current bitstream yet
-         *
-         * @return true if it has sent a key frame from this bitstream, false
-         * otherwise
-         */
-        private boolean haveSentKeyFrame()
-        {
-            return mostRecentSentKeyFrame != null;
-        }
-
-        /**
-         * Returns whether or not this bitstream is suspended
-         *
-         * @return true if this bitstream is suspended, false otherwise
-         */
-        private boolean isSuspended()
-        {
-            return currentIdx == -1;
-        }
-
-        /**
-         * Utility class that holds information about seen frames.
-         */
-        class SeenFrame
-        {
-            /**
-             * The sequence number translation to apply to accepted RTP packets.
-             */
-            private SeqNumTranslation seqNumTranslation;
-
-            /**
-             * The RTP timestamp translation to apply to accepted RTP/RTCP packets.
-             */
-            private TimestampTranslation tsTranslation;
-
-            /**
-             * A boolean that indicates whether or not the transform thread should
-             * try to piggyback missed packets from the initial key frame.
-             */
-            private boolean maybeFixInitialIndependentFrame = true;
-
-            /**
-             * The source timestamp of this frame.
-             */
-            private long srcTs;
-
-            /**
-             * The VP8 picture id to set to outgoing packets that belong to this
-             * frame.
-             */
-            private int dstPictureID = -1;
-
-            /**
-             * The VP8 TL0PICIDX to set to outgoing packets that belong to this
-             * frame.
-             * https://tools.ietf.org/html/rfc7741#section-4.2
-             */
-            private int dstTL0PICIDX = -1;
-
-            /**
-             * The maximum source sequence number to accept. -1 means drop.
-             */
-            private int srcSeqNumLimit = -1;
-
-            /**
-             * The start source sequence number of this frame.
-             */
-            private int srcSeqNumStart = -1;
-
-            /**
-             * Defines an RTP packet filter that controls which packets to be
-             * written into some arbitrary target/receiver that owns this
-             * {@link SeenFrame}.
-             *
-             * @param source the {@link FrameDesc} that the RTP packet belongs
-             * to.
-             * @param pkt the packet for which to decide to accept
-             * @return <tt>true</tt> to allow the specified packet/<tt>buffer</tt>
-             * to be written into the arbitrary target/receiver that owns this
-             * {@link SeenFrame} ; otherwise, <tt>false</tt>
-             */
-            boolean accept(FrameDesc source, RawPacket pkt)
-            {
-                if (this == mostRecentSentFrame /* the max frame can expand */
-                    || availableQualityIndices == null
-                    || availableQualityIndices.length < 2)
-                {
-                    int end = source.getEnd();
-                    srcSeqNumLimit
-                        = end != -1 ? source.getEnd() : source.getMaxSeen() ;
-                }
-
-                if (srcSeqNumLimit == -1)
-                {
-                    return false;
-                }
-
-                int start = source.getStart();
-                srcSeqNumStart
-                    = start != -1 ? source.getStart() : source.getMinSeen();
-
-                int seqNum = pkt.getSequenceNumber();
-
-                // We'll accept the packet if it's no newer than our srcSeqNumLimit
-                // FIXME(brian): in what scenario will this test fail?  mostRecentSentFrame
-                // will always, at least, be equal to the frame to which this
-                // sequence number belongs, and we update srcSeqNumLimit based
-                // on the highest seen sequence number of that frame.  if seqNum
-                // was from a new frame
-                boolean accept =
-                    RTPUtils.isOlderSequenceNumberThan(
-                        seqNum,
-                        RTPUtils.applySequenceNumberDelta(srcSeqNumLimit, 1));
-
-                if (!accept && logger.isDebugEnabled())
-                {
-                    logger.debug("frame_corruption seq=" + seqNum
-                        + ",seq_start=" + srcSeqNumStart
-                        + ",seq_limit=" + srcSeqNumLimit);
-                }
-
-                return accept;
-            }
-
-            /**
-             * Rewrite the vp8 picture id
-             * @param pktOut the packet in which to rewrite the picture id
-             * @return a RawPacket matching the given one in all ways but
-             * with the picture id appropriately rewritten (may not be the same
-             * RawPacket instance as pktOut)
-             */
-            private RawPacket handleVp8PictureIdRewriting(RawPacket pktOut)
-            {
-                MediaStreamTrackDesc source = weakSource.get();
-                assert source != null;
-
-                REDBlock redBlock = source.getMediaStreamTrackReceiver()
-                    .getStream().getPrimaryREDBlock(pktOut);
-
-                if (!DePacketizer
-                    .VP8PayloadDescriptor.hasExtendedPictureId(
-                        redBlock.getBuffer(), redBlock.getOffset(),
-                        redBlock.getLength()))
-                {
-                    // XXX we have observed that using a non-extended
-                    // picture ID makes the Chrome 58 jitter buffer to
-                    // completely freak out. So here we expand the non
-                    // extended picture id field and convert it to an
-                    // extended one.
-                    byte[] srcBuf = pktOut.getBuffer();
-                    byte[] dstBuf = new byte[srcBuf.length + 1];
-                    System.arraycopy(
-                        srcBuf, 0, dstBuf, 0, redBlock.getOffset() + 3);
-                    System.arraycopy(srcBuf, redBlock.getOffset() + 3,
-                        dstBuf, redBlock.getOffset() + 4,
-                        srcBuf.length - redBlock.getOffset() - 3);
-
-                    // set the extended picture id bit.
-                    dstBuf[redBlock.getOffset() + 2] |= (byte) (0x80);
-
-                    pktOut = new RawPacket(dstBuf,
-                        pktOut.getOffset(), pktOut.getLength() + 1);
-
-                    logger.debug("Extending the picture ID of a VP8 pkt.");
-                }
-
-                if (!DePacketizer
-                    .VP8PayloadDescriptor.setExtendedPictureId(
-                        // XXX pktOut is not a typo.
-                        pktOut.getBuffer(), redBlock.getOffset(),
-                        redBlock.getLength(), dstPictureID))
-                {
-                    logger.warn("Failed to set the VP8 extended" +
-                        " picture ID.");
-                }
-                // We may have re-assigned pktOut to a new RawPacket
-                return pktOut;
-            }
-
-            /**
-             * Handles setting the appropriate TL0PICIDX on the given packet
-             * @param pktOut the packet in which to set the TL0PICIDX (in place)
-             *
-             */
-            private void handleSettingVp8TL0PICIDX(RawPacket pktOut)
-            {
-                MediaStreamTrackDesc source = weakSource.get();
-                assert source != null;
-
-                REDBlock redBlock = source.getMediaStreamTrackReceiver()
-                    .getStream().getPrimaryREDBlock(pktOut);
-
-                if (!DePacketizer.VP8PayloadDescriptor.setTL0PICIDX(
-                    redBlock.getBuffer(), redBlock.getOffset(),
-                    redBlock.getLength(), dstTL0PICIDX))
-                {
-                    logger.warn("Failed to set the VP8 TL0PICIDX.");
-                }
-            }
-
-            /**
-             * Translates accepted packets and drops packets that are not
-             * accepted (by this instance).
-             *
-             * @param pktIn the packet to translate or drop.
-             *
-             * @return the translated packet (along with any piggy backed
-             * packets), if the packet is accepted, null otherwise.
-             */
-            RawPacket[] rtpTransform(RawPacket pktIn)
-            {
-                RawPacket[] pktsOut;
-                long srcSsrc = pktIn.getSSRCAsLong();
-
-                if (maybeFixInitialIndependentFrame)
-                {
-                    maybeFixInitialIndependentFrame = false;
-
-                    if (srcSeqNumStart != pktIn.getSequenceNumber())
-                    {
-                        MediaStreamTrackDesc source = weakSource.get();
-                        assert source != null;
-                        // Piggy back till max seen.
-                        RawPacketCache inCache = source
-                            .getMediaStreamTrackReceiver()
-                            .getStream()
-                            .getCachingTransformer()
-                            .getIncomingRawPacketCache();
-
-                        int len = RTPUtils.getSequenceNumberDelta(
-                            srcSeqNumLimit, srcSeqNumStart) + 1;
-                        pktsOut = new RawPacket[len];
-                        for (int i = 0; i < pktsOut.length; i++)
-                        {
-                            // Note that the ingress cache might not have the desired
-                            // packet.
-                            pktsOut[i] = inCache.get(srcSsrc, (srcSeqNumStart + i) & 0xFFFF);
-                        }
-                    }
-                    else
-                    {
-                        pktsOut = new RawPacket[] { pktIn };
-                    }
-                }
-                else
-                {
-                    pktsOut = new RawPacket[]{ pktIn };
-                }
-
-                for (int i = 0; i < pktsOut.length; i++)
-                {
-                    RawPacket pktOut = pktsOut[i];
-
-                    // Note that the ingress cache might not have the desired packet.
-                    if (pktOut == null)
-                    {
-                        continue;
-                    }
-
-                    if (seqNumTranslation != null)
-                    {
-                        int srcSeqNum = pktOut.getSequenceNumber();
-                        int dstSeqNum = seqNumTranslation.apply(srcSeqNum);
-
-                        if (srcSeqNum != dstSeqNum)
-                        {
-                            pktOut.setSequenceNumber(dstSeqNum);
-                        }
-                    }
-
-                    if (tsTranslation != null)
-                    {
-                        long dstTs = tsTranslation.apply(srcTs);
-
-                        if (dstTs != srcTs)
-                        {
-                            pktOut.setTimestamp(dstTs);
-                        }
-                    }
-
-                    if (ENABLE_VP8_PICID_REWRITING && dstPictureID > -1)
-                    {
-                        pktsOut[i] = handleVp8PictureIdRewriting(pktOut);
-                    }
-
-                    if (dstTL0PICIDX > -1)
-                    {
-                        handleSettingVp8TL0PICIDX(pktOut);
-                    }
-                }
-                return pktsOut;
-            }
-
-            /**
-             * Gets the max sequence number that has been sent out by this instance.
-             *
-             * @return the max sequence number that has been sent out by this
-             * instance.
-             */
-            int getMaxSeqNum()
-            {
-                return seqNumTranslation == null
-                    ? srcSeqNumLimit : seqNumTranslation.apply(srcSeqNumLimit);
-            }
-
-            /**
-             * Gets the max timestamp that has been sent out by this instance.
-             *
-             * @return the max timestamp that has been sent out by this instance.
-             */
-            long getTs()
-            {
-                return tsTranslation == null ? srcTs : tsTranslation.apply(srcTs);
-            }
-
-            /**
-             * Resets this seen frame.
-             *
-             * @param srcTs the timestamp of the seen frame.
-             * @param seqNumTranslation the {@link SeqNumTranslation} to apply to
-             * RTP packets of this frame.
-             * @param tsTranslation the {@link TimestampTranslation} to apply to
-             * RTP packets of this frame.
-             * @param dstPictureID The VP8 picture id to set to outgoing packets
-             * that belong to this frame.
-             * @param dstTL0PICIDX The VP8 TL0PICIDX to set to outgoing packets
-             * that belong to this frame.
-             */
-            void reset(
-                long srcTs,
-                SeqNumTranslation seqNumTranslation,
-                TimestampTranslation tsTranslation,
-                int dstPictureID, int dstTL0PICIDX,
-                boolean maybeFixInitialIndependentFrame)
-            {
-                this.maybeFixInitialIndependentFrame = maybeFixInitialIndependentFrame;
-                this.srcSeqNumLimit = -1;
-                this.srcSeqNumStart = -1;
-                this.srcTs = srcTs;
-                this.seqNumTranslation = seqNumTranslation;
-                this.tsTranslation = tsTranslation;
-                this.dstPictureID = dstPictureID;
-                this.dstTL0PICIDX = dstTL0PICIDX;
-            }
-
-            /**
-             * Resets this seen frame.
-             *
-             * @param srcTs the timestamp of the seen frame.
-             */
-            void reset(long srcTs)
-            {
-                reset(srcTs, null, null, -1, -1, false);
-            }
-        }
-
-        /**
-         * A very simple {@link SeenFrame} allocator. NOT multi-thread safe.
-         */
-        class SeenFrameAllocator
-        {
-            /**
-             * The pool of available {@link SeenFrame}.
-             */
-            private Queue<SeenFrame> pool = new LinkedList<>();
-
-            /**
-             * Gets a {@link SeenFrame} from the pool of available seen frames,
-             * or creates a new one, if none is available.
-             *
-             * @return a {@link SeenFrame} from the pool of available seen
-             * frames, or a new one, if none is available.
-             */
-            SeenFrame getOrCreate()
-            {
-                return pool.isEmpty() ? new SeenFrame() : pool.remove();
-            }
-
-            /**
-             * Returns a {@link SeenFrame} to this allocator.
-             *
-             * @param value the {@link SeenFrame} to return.
-             */
-            void returnSeenFrame(SeenFrame value)
-            {
-                pool.add(value);
-            }
-        }
+        bitstreamController.suspend();
     }
 }
